This function extracts key editorial information from one or more paper URLs. Specifically, it retrieves the submission, revision, and acceptance dates, as well as the article type. The function also calculates the turnaround time (the duration from submission to acceptance) and identifies whether the paper is part of a special issue.

This function extracts key editorial information from one or more paper URLs. Specifically, it retrieves the submission, revision, and acceptance dates, as well as the article type. The function also calculates the turnaround time (the duration from submission to acceptance) and identifies whether the paper is part of a special issue.

Usage

article_info(vector, sleep = 2, sample_size, show_progress = TRUE)

Arguments

vector

A vector with urls.

sleep

Number of seconds between scraping iterations. 2 sec. by default

sample_size

A number. How many papers do you want to explore from the main vector. Leave blank for all

show_progress

Logical. If TRUE, a progress bar is displayed during the function execution. Defaults to TRUE.

Value

A data frame (class: data.frame) with the following columns:

i

The URL of the article from which the information is retrieved.

article_type

The classification of the article (e.g., editorial, review).

Received

The date the article was received by the publisher.

Accepted

The date the article was accepted for publication.

tat

The turnaround time, calculated as the number of days between the received and accepted dates.

year

The year in which the article was accepted for publication.

issue_type

Indicates whether the article is part of a special issue.

Examples

url<-c("https://www.mdpi.com/2073-4336/8/4/45","https://www.mdpi.com/2073-4336/11/3/39")
# \donttest{
info<-article_info(url, 1.5)
#>   |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================================================| 100%
# }