Retrieves all special issues of a specified journal with URLs. Filters results by issue status (open, closed, or all) and optional year range.

Retrieves all special issues of a specified journal with URLs. Filters results by issue status (open, closed, or all) and optional year range.

Usage

special_issue_find(journal, type = "closed", years = NULL, verbose = TRUE)

Arguments

journal

MDPI journal code

type

"closed", "open" or "all" special issues. "closed" by default.

years

A vector containing special issues closure dates to limit the search to certain years

verbose

Logical. If TRUE, informative messages will be printed during the function execution. Defaults to TRUE.

Value

A vector.

Examples

# \donttest{
special_issue_find("covid")
#> Compiling all special issues from journal covid with status: closed
#> Extracting page: 1
#> Extracting page: 2
#> [1] "https://www.mdpi.com/journal/covid/special_issues/genetic_sars"
#> [2] "https://www.mdpi.com/journal/covid/special_issues/CVD_covid"   
#> [3] "https://www.mdpi.com/journal/covid/special_issues/FXV4HILD59"  
#> [4] "https://www.mdpi.com/journal/covid/special_issues/19W5063G34"  
#> [5] "https://www.mdpi.com/journal/covid/special_issues/4L613D592K"  
#> [6] "https://www.mdpi.com/journal/covid/special_issues/SE59O06T6Z"  
#> [7] "https://www.mdpi.com/journal/covid/special_issues/7U2577X558"  
# }